requirements.txt
APScheduler>=3.0.4
django>=1.4.0
html5validator>=0.1.14
python-slugify>=1.1.4
